Add UpdateQuestion route with position update

+++ quiz-api/MyDatabase.py
@@ -19,6 +19,11 @@ class MyDatabase():
         self.cur.execute("commit")
 
         # IF ERROR THEN cur.execute('rollback')
+    def getNbQuestions(self):
+        cur = self.connection.cursor()
+        cur.execute(
+            "Select Count(id) from questions")
+        return cur.fetchone()[0]
 
     def getParticipationResults(self):
         cur = self.connection.cursor()
@@ -72,6 +77,8 @@ class MyDatabase():
     def addQuestion(self, q: Question):
         cur = self.connection.cursor()
         cur.execute("Begin")
+        # Add 1 to position to all questions to which their position is equal or greater than q.position
+        self.updateQuestionPosition(1, q.position)
         cur.execute(
             "Insert into questions (title, image, position, text) values (?, ?, ?, ?)",
             (q.title, q.image, q.position, q.text))
@@ -80,16 +87,51 @@ class MyDatabase():
             "Insert into possible_answers (text, isCorrect, nbSips, questionId) values (?, ?, ?, ?)", [(pa.text, pa.isCorrect, pa.nbSips, q.id) for pa in q.possibleAnswers])
         cur.execute('Commit')
 
+    def updateQuestion(self, q: Question):
+        cur = self.connection.cursor()
+        cur.execute(f"Select position from questions where id = {q.id}")
+        res = cur.fetchone()
+        if res is None:
+            return False
+        currentPos = res[0]
+
+        cur.execute("Begin")
+        newPos = q.position
+        if currentPos >= newPos:
+            self.updateQuestionPosition(1, newPos, currentPos)
+        else:
+            self.updateQuestionPosition(-1, currentPos, newPos)
+
+        cur.execute(
+            "Update questions Set title=?, image=?, position=?, text=? Where id=?",
+            (q.title, q.image, q.position, q.text, q.id))
+        # Delete possible_answers of the question as we don't have their ids to directly update them
+        cur.execute(f"Delete from possible_answers Where questionId={q.id}")
+        # Add new possible_answers
+        cur.executemany(
+            "Insert into possible_answers (text, isCorrect, nbSips, questionId) values (?, ?, ?, ?)", [(pa.text, pa.isCorrect, pa.nbSips, q.id) for pa in q.possibleAnswers])
+
+        cur.execute('Commit')
+        return True
+
     def deleteQuestion(self, questionId: int):
         cur = self.connection.cursor()
+        cur.execute(f"Select position from questions where id = {questionId}")
+        res = cur.fetchone()
+        if res is None:
+            return False
+
+        position = res[0]
+
         cur.execute("Begin")
         cur.execute(
             f"Delete from possible_answers Where questionId = {questionId}")
         cur.execute(
             f"Delete from questions Where id = {questionId}")
-        rowCount = cur.rowcount
+        # Decrement 1 to position to all questions to which their position is equal or greater than q.position
+        self.updateQuestionPosition(-1, position)
         cur.execute('Commit')
-        return rowCount == 1
+        return True
 
     def deleteAllQuestions(self):
         cur = self.connection.cursor()
@@ -103,3 +145,11 @@ class MyDatabase():
         cur.execute("Begin")
         cur.execute("Delete From participation_results")
         cur.execute('Commit')
+
+    # increment : 1 or -1
+    def updateQuestionPosition(self, increment: int, startAtPos: int, endAtPos: int = -1):
+        cur = self.connection.cursor()
+        sqlRequest = f"Update questions set position = position + {increment} Where position >= {startAtPos}"
+        if endAtPos != -1:
+            sqlRequest += f" and position <= {endAtPos}"
+        cur.execute(sqlRequest)
